# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Gestión de Pedidos de Aberturas",
        "description": "API REST para gestionar los pedidos de una empresa de aberturas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Listar categorías",
                "description": "Devuelve las categorías activas ordenadas por nombre. No requiere token.",
                "operationId": "ef13f6e372e5ab339e1180104b338809",
                "responses": {
                    "200": {
                        "description": "Categorías obtenidas correctamente."
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Crear categoría",
                "description": "Crea una categoría nueva. Solo administradores.",
                "operationId": "897e4608beae455e11f07bf62c0795b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mamparas",
                                        "maxLength": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada correctamente."
                    },
                    "401": {
                        "description": "No estás autenticado."
                    },
                    "403": {
                        "description": "No tenés permiso para realizar esta acción."
                    },
                    "422": {
                        "description": "Los datos proporcionados no son válidos."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/categories/{id}": {
            "put": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Editar categoría",
                "description": "Cambia el nombre de una categoría o la vuelve a activar. Solo administradores.",
                "operationId": "50204f4a6e2b8d4cfb73679514fa1f6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ventanas corredizas",
                                        "maxLength": 100
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada correctamente."
                    },
                    "401": {
                        "description": "No estás autenticado."
                    },
                    "403": {
                        "description": "No tenés permiso para realizar esta acción."
                    },
                    "404": {
                        "description": "Categoría no encontrada."
                    },
                    "422": {
                        "description": "Los datos proporcionados no son válidos."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Desactivar categoría",
                "description": "Desactiva una categoría (no la borra). Solo administradores.",
                "operationId": "7f6d0695642fc7ec4349a20d0f447037",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría desactivada correctamente."
                    },
                    "401": {
                        "description": "No estás autenticado."
                    },
                    "403": {
                        "description": "No tenés permiso para realizar esta acción."
                    },
                    "404": {
                        "description": "Categoría no encontrada."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación generado por Laravel Sanctum.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}